# docs/openapi.yaml
---
openapi: 3.1.0
info:
  title: Bank API
  description: API for managing bank cards
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - JWT: []
tags:
  - name: Users
    description: Endpoints for user profile and administration
  - name: Authentication
    description: Endpoints for user registration, login, and token management.
  - name: Cards
    description: Endpoints for managing bank cards and transfers
paths:
  "/user/create":
    post:
      tags:
        - Users
      summary: 'Admin: Create user'
      description: Admin creates a new user with specific roles.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserCreateRequest"
        required: true
      responses:
        '201':
          description: User created successfully
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '400':
          description: Validation failed
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '403':
          description: 'Access denied: Admin role required'
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '409':
          description: 'Conflict: Username already exists'
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
  "/cards/create":
    post:
      tags:
        - Cards
      summary: 'Admin: Create card'
      description: Admin creates a new card for a specific user.
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CardCreateRequest"
        required: true
      responses:
        '201':
          description: Card created successfully
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardMaskedResponse"
        '400':
          description: Invalid request data (negative balance, etc.)
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardMaskedResponse"
        '404':
          description: Target user not found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardMaskedResponse"
  "/auth/signup":
    post:
      tags:
        - Authentication
      summary: User Registration
      description: Creates a new user account.
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SignUpRequest"
        required: true
      responses:
        '201':
          description: User successfully registered
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '400':
          description: Validation error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '409':
          description: User with this username already exists
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
  "/auth/signin":
    post:
      tags:
        - Authentication
      summary: User Login
      description: Authenticates user and returns access and refresh tokens.
      operationId: authenticateUser
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SignInRequest"
        required: true
      responses:
        '200':
          description: Successfully authenticated
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/JwtResponse"
        '400':
          description: Invalid request format
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/JwtResponse"
        '401':
          description: Invalid username or password
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/JwtResponse"
  "/auth/refresh":
    post:
      tags:
        - Authentication
      summary: Refresh Token
      description: Generates a new access token using a valid refresh token.
      operationId: refreshAccessToken
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        '200':
          description: Tokens successfully refreshed
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/JwtResponse"
        '401':
          description: Invalid or expired refresh token
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/JwtResponse"
  "/user/update/{id}":
    patch:
      tags:
        - Users
      summary: 'Admin: Update user'
      description: Admin updates data of any user by ID.
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserCreateRequest"
        required: true
      responses:
        '200':
          description: User updated successfully
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '400':
          description: Invalid input data
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '403':
          description: 'Access denied: Admin role required'
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '404':
          description: User with given ID not found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
  "/user/profile":
    delete:
      tags:
        - Users
      summary: Delete current profile
      description: Allows authorized user to delete their own account.
      operationId: deleteCurrentUser
      responses:
        '200':
          description: Account deleted successfully
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '401':
          description: Full authentication is required
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '404':
          description: Current user not found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
    patch:
      tags:
        - Users
      summary: Update current profile
      description: Allows authorized user to update their own profile data.
      operationId: updateCurrentUser
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SignUpRequest"
        required: true
      responses:
        '200':
          description: Profile updated successfully
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '400':
          description: Invalid input data
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '401':
          description: Full authentication is required
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '404':
          description: Current user not found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
  "/cards/transfer":
    patch:
      tags:
        - Cards
      summary: Transfer between own cards
      description: Transfers money between two cards belonging to the same user.
      operationId: transferBetweenOwnCards
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TransferRequest"
        required: true
      responses:
        '200':
          description: Transfer completed successfully
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TransferResponse"
        '400':
          description: Invalid input or validation error
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TransferResponse"
        '409':
          description: Incorrect PIN, same card numbers, or card is blocked
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TransferResponse"
        '422':
          description: Insufficient funds for the operation
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/TransferResponse"
  "/cards/block_request":
    patch:
      tags:
        - Cards
      summary: Request card block
      description: Allows user to request a block for their own card.
      operationId: requestToBlockCard
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CardNumberRequest"
        required: true
      responses:
        '201':
          description: Block request successfully created
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardMaskedResponse"
        '404':
          description: Card not found or doesn't belong to the user
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardMaskedResponse"
  "/cards/block":
    patch:
      tags:
        - Cards
      summary: 'Admin: Block card'
      description: Admin confirms a card block request.
      operationId: blockCard
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CardNumberRequest"
        required: true
      responses:
        '200':
          description: Card successfully blocked by admin
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardMaskedResponse"
        '403':
          description: 'Access denied: Admin role required'
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardMaskedResponse"
        '404':
          description: Block request not found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardMaskedResponse"
  "/cards/activate":
    patch:
      tags:
        - Cards
      summary: 'Admin: Activate card'
      description: Admin confirms a card activation request.
      operationId: activateCard
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CardNumberRequest"
        required: true
      responses:
        '200':
          description: Card successfully activated
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardMaskedResponse"
        '403':
          description: 'Access denied: Admin role required'
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardMaskedResponse"
        '404':
          description: Card or activation request not found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardMaskedResponse"
  "/user/get_all":
    get:
      tags:
        - Users
      summary: 'Admin: Get all users'
      description: Returns a paginated list of all registered users.
      operationId: getAll
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/Pageable"
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PagedModel"
        '403':
          description: 'Access denied: Admin role required'
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PagedModel"
  "/user/get/{id}":
    get:
      tags:
        - Users
      summary: 'Admin: Get user by ID'
      description: Admin retrieves full information about a user.
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User data retrieved successfully
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '403':
          description: 'Access denied: Admin role required'
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '404':
          description: User with given ID not found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UserResponse"
  "/cards/my_cards":
    get:
      tags:
        - Cards
      summary: Get current user cards
      description: Returns a paginated list of cards belonging to the authorized user
        (masked).
      operationId: getUserCards
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/Pageable"
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CardsFilter"
      responses:
        '200':
          description: List of masked cards retrieved successfully
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PagedModel"
        '401':
          description: Full authentication is required to access this resource
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PagedModel"
  "/cards/get_details":
    get:
      tags:
        - Cards
      summary: Get full card details
      description: Returns full card details after PIN verification. Handles failed
        attempts.
      operationId: getCardDetails
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CardPasswordRequest"
        required: true
      responses:
        '200':
          description: PIN verified, full details returned
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardResponse"
        '401':
          description: Incorrect PIN code provided
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardResponse"
        '404':
          description: Card not found
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardResponse"
        '409':
          description: Card is blocked or too many failed attempts
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CardResponse"
  "/cards/get_all":
    get:
      tags:
        - Cards
      summary: 'Admin: Get all cards'
      description: Returns a paginated list of all cards in the system (full data).
      operationId: getAllCards
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/Pageable"
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CardsFilter"
      responses:
        '200':
          description: Full list of cards retrieved successfully
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PagedModel"
        '403':
          description: 'Access denied: Admin role required'
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PagedModel"
  "/user/delete/{id}":
    delete:
      tags:
        - Users
      summary: 'Admin: Delete user'
      description: Admin deletes any user by ID.
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User deleted successfully
          content:
            "*/*":
              schema:
                type: object
        '403':
          description: 'Access denied: Admin role required'
          content:
            "*/*":
              schema:
                type: object
        '404':
          description: User with given ID not found
          content:
            "*/*":
              schema:
                type: object
  "/cards/delete/{id}":
    delete:
      tags:
        - Cards
      summary: 'Admin: Delete card'
      description: Permanently deletes a card by ID.
      operationId: deleteCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Card deleted successfully
          content:
            "*/*":
              schema:
                type: object
        '403':
          description: 'Access denied: Admin role required'
          content:
            "*/*":
              schema:
                type: object
        '404':
          description: Card ID not found
          content:
            "*/*":
              schema:
                type: object
components:
  schemas:
    UserCreateRequest:
      type: object
      properties:
        username:
          type: string
        lastName:
          type: string
        firstName:
          type: string
        secondName:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
        password:
          type: string
          maxLength: 2147483647
          minLength: 6
      required:
        - firstName
        - lastName
        - password
        - role
        - secondName
        - username
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        FIO:
          type: string
    CardCreateRequest:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        balance:
          type: number
        pincode:
          type: string
          maxLength: 4
          minLength: 4
          pattern: "\\d+"
      required:
        - balance
        - pincode
        - userId
    CardMaskedResponse:
      type: object
      properties:
        maskedNumber:
          type: string
        status:
          type: string
        ownerName:
          type: string
        expiryDate:
          type: string
          format: date
    SignUpRequest:
      type: object
      properties:
        username:
          type: string
        lastName:
          type: string
        firstName:
          type: string
        secondName:
          type: string
        password:
          type: string
          maxLength: 2147483647
          minLength: 6
        confirmPassword:
          type: string
        passwordConfirmed:
          type: boolean
      required:
        - confirmPassword
        - firstName
        - lastName
        - password
        - secondName
        - username
    SignInRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          maxLength: 2147483647
          minLength: 6
      required:
        - password
        - username
    JwtResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        type:
          type: string
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    TransferRequest:
      type: object
      properties:
        fromNumber:
          type: string
          maxLength: 16
          minLength: 16
          pattern: "\\d+"
        toNumber:
          type: string
          maxLength: 16
          minLength: 16
          pattern: "\\d+"
        amount:
          type: number
        pincode:
          type: string
          maxLength: 4
          minLength: 4
          pattern: "\\d+"
      required:
        - amount
        - fromNumber
        - pincode
        - toNumber
    TransferResponse:
      type: object
      properties:
        message:
          type: string
    CardNumberRequest:
      type: object
      properties:
        number:
          type: string
          maxLength: 16
          minLength: 16
          pattern: "\\d+"
      required:
        - number
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModel:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          "$ref": "#/components/schemas/PageMetadata"
    CardsFilter:
      type: object
      properties:
        status:
          type: string
        date:
          type: string
          format: date
    CardPasswordRequest:
      type: object
      properties:
        lastFourDigitsOfNumber:
          type: string
          maxLength: 4
          minLength: 4
          pattern: "\\d+"
        pincode:
          type: string
          maxLength: 4
          minLength: 4
          pattern: "\\d+"
      required:
        - lastFourDigitsOfNumber
        - pincode
    CardResponse:
      type: object
      properties:
        number:
          type: string
        status:
          type: string
        ownerName:
          type: string
        balance:
          type: number
        expiryDate:
          type: string
          format: date
  securitySchemes:
    JWT:
      type: http
      description: Insert your Access Token here
      scheme: bearer
      bearerFormat: JWT