# src/main/java/com/example/bankcards/dto/card/request/CardsFilter.java
package com.example.bankcards.dto.card.request;

import lombok.Getter;
import org.springframework.format.annotation.DateTimeFormat;

import java.time.LocalDate;

@Getter
public class CardsFilter {
    private String status;
    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
    private LocalDate date;
}